anual: finish cards

+++ anual/src/main/java/org/grupo11/Services/Card/CardUsage.java
@@ -4,9 +4,9 @@ import org.grupo11.Services.Fridge.Fridge;
 
 public class CardUsage {
     Fridge fridge;
-    int usedAt;
+    long usedAt;
 
-    public CardUsage(Fridge fridge, int usedAt) {
+    public CardUsage(Fridge fridge, long usedAt) {
         this.fridge = fridge;
         this.usedAt = usedAt;
     }
